+++ NEWS.md
@@ -1,6 +1,7 @@
 # inlmisc 0.4.5.9000
 
-- ...
+- In `AddSearchButton` function, fix bug that prevented popup from opening,
+  occurred when `openPopup` argument is true.
 
 # inlmisc 0.4.5