+++ NEWS.md
@@ -1,3 +1,10 @@
+# inlmisc 0.4.5.9000
+
+- In `PrintFigure` function, allow heading on single figure.
+
+- In `AddSearchButton` function, fix bug that prevented popup from opening,
+  occurred when `openPopup` argument is true.
+
 # inlmisc 0.4.5
 
 - In `PlotGraph` function, change alignment of `main` title from center to left-hand side of plot.